feat: Implement core platform features including quest management,       user profiles, and skill progression part 2
