+++ app/ai-rank-test/results/page.tsx
@@ -8,7 +8,6 @@ import { Badge } from '@/components/ui/badge'
 import { Progress } from '@/components/ui/progress'
 import { Alert, AlertDescription } from '@/components/ui/alert'
 import {
-  Trophy,
   Star,
   TrendingUp,
   Target,
@@ -163,7 +162,7 @@ const rankDetails = {
 
 export default function AIRankTestResults() {
   const router = useRouter()
-  const { profile } = useAuth()
+  useAuth() // Note: useAuth hook is called but its return value (profile) is not used.
   const [isLoading, setIsLoading] = useState(true)
   const [animatedScore, setAnimatedScore] = useState(0)
   const [showDetails, setShowDetails] = useState(false)
