+++ app/api/quest_applications/[id]/status/route.ts
@@ -1,5 +1,5 @@
 
-import { createClient } from '@/lib/supabase/server'
+
 import { NextRequest, NextResponse } from 'next/server'
 import { z } from 'zod'
 
@@ -9,7 +9,7 @@ const statusSchema = z.object({
 
 export async function PUT(req: NextRequest, { params }: { params: { id: string } }) {
   try {
-    const supabase = createClient()
+    const supabase = createSupabaseServerClient()
     const { id: applicationId } = params
     const body = await req.json()
