+++ app/api/quest_submissions/[id]/status/route.ts
@@ -1,5 +1,5 @@
 
-import { createClient } from '@/lib/supabase/server'
+
 import { NextRequest, NextResponse } from 'next/server'
 import { z } from 'zod'
 
@@ -10,7 +10,7 @@ const statusSchema = z.object({
 
 export async function PUT(req: NextRequest, { params }: { params: { id: string } }) {
   try {
-    const supabase = createClient()
+    const supabase = createSupabaseServerClient()
     const { id: submissionId } = params
     const body = await req.json()
