+++ app/api/quests/[id]/applications/route.ts
@@ -1,10 +1,10 @@
 
-import { createClient } from '@/lib/supabase/server'
+
 import { NextRequest, NextResponse } from 'next/server'
 
 export async function GET(req: NextRequest, { params }: { params: { id: string } }) {
   try {
-    const supabase = createClient()
+    const supabase = createSupabaseServerClient()
     const { id: quest_id } = params
 
     const { data: applications, error } = await supabase
