+++ app/api/quests/[id]/submit/route.ts
@@ -1,5 +1,5 @@
 
-import { createClient } from '@/lib/supabase/server'
+
 import { NextRequest, NextResponse } from 'next/server'
 import { z } from 'zod'
 
@@ -14,7 +14,7 @@ const submissionSchema = z.object({
 
 export async function POST(req: NextRequest, { params }: { params: { id: string } }) {
   try {
-    const supabase = createClient()
+    const supabase = createSupabaseServerClient()
     const { id: quest_id } = params
     const body = await req.json()
