+++ package.json
@@ -79,9 +79,11 @@
     "@types/node": "^22",
     "@types/react": "^19",
     "@types/react-dom": "^19",
+    "eslint": "^9.34.0",
+    "eslint-config-next": "15.5.0",
     "postcss": "^8.5",
     "supabase": "^2.34.3",
     "tailwindcss": "^3.4.17",
-    "typescript": "^5"
+    "typescript": "^5.9.2"
   }
 }
